chore: refactor events table
