+++ web/src/features/events/hooks/useEventsFilterOptions.ts
@@ -1,0 +1,86 @@
+import { api } from "@/src/utils/api";
+import { useMemo } from "react";
+import { type FilterState, type TimeFilter } from "@langfuse/shared";
+
+type UseEventsFilterOptionsParams = {
+  projectId: string;
+  oldFilterState: FilterState;
+};
+
+export function useEventsFilterOptions({
+  projectId,
+  oldFilterState,
+}: UseEventsFilterOptionsParams) {
+  // Extract start time filters for filter options query
+  const startTimeFilters = useMemo(() => {
+    return oldFilterState.filter(
+      (f) =>
+        (f.column === "Start Time" || f.column === "startTime") &&
+        f.type === "datetime",
+    ) as TimeFilter[];
+  }, [oldFilterState]);
+
+  // Fetch filter options
+  const filterOptions = api.events.filterOptions.useQuery(
+    {
+      projectId,
+      startTimeFilter:
+        startTimeFilters.length > 0 ? startTimeFilters : undefined,
+    },
+    {
+      trpc: {
+        context: {
+          skipBatch: true,
+        },
+      },
+      refetchOnMount: false,
+      refetchOnWindowFocus: false,
+      refetchOnReconnect: false,
+      staleTime: Infinity,
+    },
+  );
+
+  // Transform filter options for sidebar
+  const newFilterOptions = useMemo(() => {
+    const scoreCategories =
+      filterOptions.data?.score_categories?.reduce(
+        (acc, score) => {
+          acc[score.label] = score.values;
+          return acc;
+        },
+        {} as Record<string, string[]>,
+      ) ?? undefined;
+
+    const scoresNumeric = filterOptions.data?.scores_avg ?? undefined;
+
+    return {
+      environment: filterOptions.data?.environment ?? undefined,
+      name: filterOptions.data?.name ?? undefined,
+      type: filterOptions.data?.type ?? undefined,
+      level: filterOptions.data?.level ?? undefined,
+      providedModelName: filterOptions.data?.providedModelName ?? undefined,
+      modelId: filterOptions.data?.modelId ?? undefined,
+      promptName: filterOptions.data?.promptName ?? undefined,
+      traceTags: filterOptions.data?.traceTags ?? undefined,
+      userId: filterOptions.data?.userId ?? undefined,
+      sessionId: filterOptions.data?.sessionId ?? undefined,
+      version: filterOptions.data?.version ?? undefined,
+      latency: [],
+      timeToFirstToken: [],
+      tokensPerSecond: [],
+      inputTokens: [],
+      outputTokens: [],
+      totalTokens: [],
+      inputCost: [],
+      outputCost: [],
+      totalCost: [],
+      score_categories: scoreCategories,
+      scores_avg: scoresNumeric,
+    };
+  }, [filterOptions.data]);
+
+  return {
+    filterOptions,
+    newFilterOptions,
+  };
+}
